replace main repl
