+++ website/src/pages/index.astro
@@ -1,11 +1,11 @@
 ---
-import HeadCommon from '../components/HeadCommon.astro';
-import { Repl } from '../repl/Repl.jsx';
+import HeadCommonNext from '../components/HeadCommonNext.astro';
+import { Repl } from '../repl/Repl';
 ---
 
 <html lang="en" class="dark">
   <head>
-    <HeadCommon />
+    <HeadCommonNext />
     <title>Strudel REPL</title>
   </head>
   <body class="h-app-height bg-background">
