--- website/src/pages/next/index.astro
@@ -1,14 +1,0 @@
----
-import HeadCommonNext from '../../components/HeadCommonNext.astro';
-import { Repl2 } from '../../repl/Repl2';
----
-
-<html lang="en" class="dark">
-  <head>
-    <HeadCommonNext />
-    <title>Strudel REPL</title>
-  </head>
-  <body class="h-app-height bg-background">
-    <Repl2 client:only="react" />
-  </body>
-</html>
